Extract bout page into its own component, update path for new bout to be slightly more RESTful (allows bouts/[id])
